requirements.txt
selenium==4.18.1
mysql-connector-python==8.3.0
# psycopg2-binary==2.9.9
pandas
sqlalchemy
matplotlib
seaborn
Sastrawi
Flask
jsonify
load_dotenv